requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
djangorestframework==3.16.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
oauthlib==3.3.1
ollama==0.6.1
pillow==12.0.0
praw==7.8.1
prawcore==2.4.0
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
tweepy==4.16.0
typing-inspection==0.4.2
typing_extensions==4.15.0
update-checker==0.18.0
urllib3==2.5.0
websocket-client==1.9.0
